refactor(v1): share CLI resume primitives
